Clear out errors once we have handled them.
In particular, in the type-hash code we can handle a failure
to encode or parse the type hash from the user data. Once we've
handled it, and printed out an error about it, we should clear
it because we are going to continue on and something else down
the line might have an error.

This will avoid errors like:

This error state is being overwritten:

with this new error message:

Signed-off-by: Chris Lalancette <clalancette@gmail.com>