Revamp discovery_options handling.
discovery_options is a per-context concept, so it should
not be set when initializing a new node.  Instead, it
should be set when the context is initialized.

Further, there is no reason to copy the discovery_options
structure into the rmw_context_impl_s; it already exists
in the rmw_context_t, which we are keeping a reference to.

This code cleans all of that up, and thus fixes the memory
leak caused by allocating this value during initialize_node
time.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>